+++ libraries/stdlib/unsigned/src/kotlin/ULong.kt
@@ -150,13 +150,13 @@ public value class ULong @PublishedApi internal constructor(@PublishedApi intern
 
     /** TODO */
     @kotlin.internal.InlineOnly
-    public inline fun mod(other: UByte): ULong = this.mod(other.toULong())
+    public inline fun mod(other: UByte): UByte = this.mod(other.toULong()).toUByte()
     /** TODO */
     @kotlin.internal.InlineOnly
-    public inline fun mod(other: UShort): ULong = this.mod(other.toULong())
+    public inline fun mod(other: UShort): UShort = this.mod(other.toULong()).toUShort()
     /** TODO */
     @kotlin.internal.InlineOnly
-    public inline fun mod(other: UInt): ULong = this.mod(other.toULong())
+    public inline fun mod(other: UInt): UInt = this.mod(other.toULong()).toUInt()
     /** TODO */
     @kotlin.internal.InlineOnly
     public inline fun mod(other: ULong): ULong = rem(other)
